(BSR)[API] refactor: move dms serializer test to corresponding folder
